requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# VGG16 and Computer Vision
tensorflow>=2.13.0
keras>=2.13.0
opencv-python>=4.8.0
Pillow>=10.0.0

# BLIP and Vision-Language Models
salesforce-lavis>=1.0.0
sentence-transformers>=2.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development and Testing
jupyter>=1.0.0
notebook>=6.5.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced features
accelerate>=0.20.0
bitsandbytes>=0.39.0
